requirements.txt
# UAgent System Dependencies

# ===== Core Framework =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ===== Async and Concurrency =====
aiohttp==3.9.1
httpx==0.25.2
asyncio-mqtt==0.16.1
websockets==12.0
python-socketio==5.10.0

# ===== Task Queue and Workflow =====
celery==5.3.4
redis==5.0.1
aioredis==2.0.1
prefect==2.14.11
python-statemachine==2.1.2
transitions==0.9.0

# ===== Database and ORM =====
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1
# asyncpg-pool==0.8.0

# ===== AI and Machine Learning =====
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.13
openai==1.6.1
anthropic==0.8.1
sentence-transformers==2.2.2
transformers==4.36.2
chromadb==0.4.18
tiktoken==0.5.2

# ===== Data Processing =====
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# ===== Text Processing and NLP =====
nltk==3.8.1
spacy==3.7.2
jinja2==3.1.2

# ===== Configuration and Settings =====
pyyaml==6.0.1
python-dotenv==1.0.0
dynaconf==3.2.4

# ===== Security and Authentication =====
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
casbin==1.17.6
authlib==1.2.1
bleach==6.1.0
validators==0.22.0

# ===== Monitoring and Logging =====
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0

# ===== Caching and Performance =====
aiocache==0.12.2
cachetools==5.3.2
psutil==5.9.6
memory-profiler==0.61.0

# ===== File and System Operations =====
aiofiles==23.2.1
watchfiles==0.21.0
pathlib2==2.3.7

# ===== Message Queue =====
aio-pika==9.3.1
aiokafka==0.8.11

# ===== Development and Testing =====
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0

# ===== Code Quality =====
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3
bandit==1.7.5
autoflake==2.2.1
pre-commit==3.6.0

# ===== Documentation =====
mkdocs==1.5.3
mkdocs-material==9.4.8

# ===== Utilities =====
click==8.1.7
rich==13.7.0
typer==0.9.0
